requirements.txt
py-cord
requests
wikipedia
google
aiohttp
asyncio
psutil
pytz
aiohttp
python-dotenv
psutil
